bugfix: ba::ui::QUI when SDLK_ESCAPE happen, _windowState_checkAll destroy locker and winState immediately, which make pollQuit crash

+++ BA_1_2/BA_Jieba.cpp
@@ -165,7 +165,7 @@ const
 	}
 
 	status.resize(X);
-	for (int x = X - 1; x >= 0; x--) {//warning C4267
+	for (int x = (int)(X - 1); x >= 0; x--) {//warning C4267
 		status[x] = stat;
 		stat = path[x + stat * X];
 	}
